Classes modified to work with ActiveRecord 3.
The version's column changes was renamed to "modifications" because AR3 does not allow
an attribute to have the name "changes" (since the Base class already has one method
with this name).
One test was also changed because it relied on the AR querying API instead of just
testing the behavior on a black box way; it was changed to do that.