FIX ordering with Versions#at with Time
When querying `#at(value)` with a time value, the result returned is the
`last` one, with no condition on ordering. Though it is the default
ordering with `id` that is used. While this is fine with incremental id,
this will fail with other types like uuids.
Fix it by explicitly telling the order clause to use the field
`created_at`